Add files via upload
This is an analytical report showing the Human resource performance.  We are to find the areas where for lay-off in comparison with the employees due for Promotion.
I connected  to excel using POWER BI to do my ETL and visualization to get insights on the data. All files for this analysis can be found here on my Repo.